LayoutTests/fast/multicol/newmulticol/balance-images.html

<!DOCTYPE html>
<html>
  <head>
    <title>balancing multicol with max-height</title>
    <script src="../../../resources/js-test-pre.js"></script>
    <script>
      if (window.testRunner)
        testRunner.dumpAsText();
      if (window.internals)
        internals.settings.setRegionBasedColumnsEnabled(true);
    </script>
    <style>
      img { display:block; height:123px; }
    </style>
  </head>
  <body>
    <p>There should be a 2 by 2 image grid below.</p>
    <div id="mc" style="-webkit-columns:3; columns:3; orphans:1; widows:1; color:olive; background:olive;">
      <img src="../resources/ipad.jpg">
      <img src="../resources/ipad.jpg">
      <img src="../resources/ipad.jpg">
      <img src="../resources/ipad.jpg">
    </div>
    <script>
      shouldBe("mc.offsetHeight", "246");
      shouldBe("document.documentElement.scrollWidth", "document.documentElement.clientWidth");
    </script>
    <script src="../../../resources/js-test-post.js"></script>
  </body>
</html>
